requirements.txt
yfinance==0.2.38
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
